base: Only add libhwjpeg to build dirs if TARGET_USES_UNIVERSAL_LIBHWJPEG is set to True
This is for the libhwjpeg version here to not interfere with different
libhwjpeg sources, and proprietary blobs, elsewhere for devices needing
them.

All current SoC configs for which the version of libhwjpeg here is used
set TARGET_USES_UNIVERSAL_LIBHWJPEG to True, so this doesn't cause any
change for them.

Change-Id: Ie569c0a5927606c678234e8456d89af66cbd680e